This is just a template website
I followed alongside the flask tutorial videos, and made this (i didn't change any of the things to make it relevent yet)

so far this is what i saw we could do with the homepage.

i think after watching a few more videos, I might try to edit one that would be our version of logans app?

This is just very basic version so far